[MIPS experimental] A new, combined method of resolving the GC pointer dereference issue.
This commit combines the original "dump the pointer in the code" idea with a minimalized version of the OBJPTR macros.

This is obviously not a final version, it will need some cleanup and rework, but it seems that the basic concept works (although this approach of sometimes using pointers and sometimes lui/ori references will surely cause similar chaos in the future).