[MIPS] Fix MinorKey layout problem in RecordWriteStub.
The total size of the bitfield in the RecordWriteStub MinorKey was
26 bits, and that exceeded the max size of CodeStub MinorKey size,
which is set by this expression in code-stubs.h:
  static const int kMinorBits = kBitsPerInt - kSmiTagSize - kMajorBits;
kMajor bits is 6, so that leaves 25 bits max for MinorKey.

This did not come up until using Crankshaft, where we do save FP regs,
thereby setting the MS bit.

This caused v8-benchmark Splay test to fail, and several Mozilla tests,
and lots of tests in debug mode.